修复石头剪刀布 bug ，启动脚本新增选择彩色输出

+++ box.sh
@@ -7,11 +7,14 @@
 # Created Time: 2020-10-03 12:48:00
 ###############################################
 
-
-dpkg -s lolcat &> /dev/null
-if [ $? -ne 0 ]; then
-    echo -e "检测到未安装 lolcat\n正在为您安装 lolcat"
-    apt install lolcat -y
+read -n 1 -p "是否需要彩色输出 (y/n): " want
+if [ "$want" = "y" ]; then
+    dpkg -s lolcat &> /dev/null
+    if [ $? -ne 0 ]; then
+        echo -e "检测到未安装 lolcat\n正在为您安装 lolcat"
+        apt install lolcat -y
+    fi
+    ./box.py | lolcat
+else
+    ./box.py | lolcat
 fi
-
-~/Box/box.py | lolcat
